requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
decorator==4.4.2
Django==4.1
djangorestframework==3.13.1
ez-setup==0.9
gunicorn==20.1.0
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.2
Pillow==9.2.0
proglog==0.1.10
psycopg2==2.9.3
python-decouple==3.6
pytube==12.1.0
pytube3==9.6.4
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
